updated again

+++ scripts/stop_container.sh
@@ -2,6 +2,7 @@
 set -e
 
 # Stop the running container (if any)
-containerid=`docker ps | awk -F " " '{print $1}'`
-docker rm -f $containerid
-
+containerid=$(docker ps -q -f ancestor=manik459/simple-python-flask-app:latest)
+if [ -n "$containerid" ]; then
+  docker rm -f $containerid
+fi
